feat: WI-0503 extract scheduling anomaly report helpers
